+++ src/components/map-controls/FullScreenControl/FullScreenControl.tsx
@@ -1,4 +1,4 @@
-import HomeVM from '@arcgis/core/widgets/Home/HomeViewModel';
+import FullScreenVM from '@arcgis/core/widgets/FullScreen/FullScreenViewModel';
 import * as React from 'react';
 
 import { useCurrentMapView, useWatchState } from '@/arcgis/hooks';
@@ -6,22 +6,19 @@ import { useCurrentMapView, useWatchState } from '@/arcgis/hooks';
 import { MapButton } from '../../Button/MapButton';
 import SvgIcon from '../../SvgIcon';
 
-function HomeControl({ viewPoint }: { viewPoint?: __esri.Viewpoint }) {
+function FullScreenControl() {
   const mapView = useCurrentMapView();
-  const widget = React.useMemo(
-    () => new HomeVM({ view: mapView, viewpoint: viewPoint }),
-    [mapView, viewPoint],
-  );
+  const widget = React.useMemo(() => new FullScreenVM({ view: mapView }), [mapView]);
   const isDisabled = useWatchState(() => widget.state === 'disabled') ?? false;
   return (
     <MapButton
       includeBorder
-      icon={<SvgIcon name="icon-home" size={14} />}
-      aria-label="Home"
+      icon={<SvgIcon name="icon-fullscreen" size={14} />}
+      aria-label="Full Screen"
       isDisabled={isDisabled}
-      onPress={() => widget.go()}
+      onPress={() => widget.toggle()}
     />
   );
 }
 
-export default HomeControl;
+export default FullScreenControl;