+++ src/routes/__root.tsx
@@ -15,6 +15,7 @@ const assetSearchSchema = z.object({
   'hide-ui': fallback(z.boolean().optional(), undefined),
   'show-region': fallback(z.boolean().optional(), undefined),
   'show-asset-popup': fallback(z.boolean().optional(), undefined),
+  'show-full-screen': fallback(z.boolean().optional(), undefined),
 });
 
 export const Route = createRootRoute({